% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jags_functions.R
\name{get_default_jags_model}
\alias{get_default_jags_model}
\title{Write a default jags model}
\usage{
get_default_jags_model(
  outcome_variable = "continuous",
  nine_point_scale_continuous = FALSE,
  residuals = "homogeneous",
  subsections = FALSE,
  path = "default_jags_model.txt",
  quiet = FALSE
)
}
\arguments{
\item{outcome_variable}{the nature of the outcome variable: continuous or
ordinal. (default = "continuous")}

\item{nine_point_scale_continuous}{only needed if nine-point continuous scale
is used (default = FALSE).}

\item{residuals}{the nature of the residuals in the model: homogeneous or
heterogeneous. (default = "homogeneous")}

\item{subsections}{the individual votes come from different sub-panels that
need to be merged. (default = FALSE)}

\item{path}{where should it be stored, from getwd()? (default = jags_model.txt)}

\item{quiet}{if TRUE, do not show messages.}
}
\description{
This function simply stores a txt file with a default jags model, in a
certain path.
}
\details{
The model defined here has a random component for the application /
the proposal and the voter / the evaluator. If the parameter subpanels is set
to TRUE an addtional grouping for the section or panel is defined. The user
is invited to write their own model definition if more flexibility is needed.
The path to the latter can then be given as parameter to get_mcmc_samples.
The user can however already decide between a continuous or ordinal outcome
variable, homogeneous or heterogeneous residuals, and with or without
subpanel merging (those options are integrated).
}
\examples{
# The model definition .txt is stored in the file "default_jags_model.txt"
\dontrun{
get_default_jags_model()
}
}
