% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_helpers.R
\name{get_inits_overdispersed_four_chains}
\alias{get_inits_overdispersed_four_chains}
\title{Helper function to define overdispersed initial values}
\usage{
get_inits_overdispersed_four_chains(
  merging_sections = FALSE,
  ordinal_scale = FALSE,
  point_scale = NULL,
  heterogeneous_residuals = FALSE,
  n_applications,
  n_voters,
  n_sections = NULL,
  grades = NULL,
  seed
)
}
\arguments{
\item{merging_sections}{should the sections be merged?, default is FALSE,
e.g. a ranking is done for each section/panel separately.}

\item{ordinal_scale}{are the evaluation scores on an ordinal scale?,
default is FALSE.}

\item{point_scale}{if on an ordinal scale, what is the number of points on
the scale?, default is NULL}

\item{heterogeneous_residuals}{dummy variable informing us on whether or not
the residuals should be heterogeneous, default = FALSE.}

\item{n_applications}{number of applications / proposals considered.}

\item{n_voters}{number of voters / panel members.}

\item{n_sections}{number of sections / panels, if merged, default is NULL.}

\item{grades}{the grades given to the proposals (the whole
n_applications*n_voters long vector)}

\item{seed}{seed used for the sampling}
}
\value{
a list of overdispersed initial values for the two chains
}
\description{
Note that this function can only be applied when four chains are used,
otherwise the user has to provide the initial values themselves. It is
however recommended to use four chain. The helper function will also use the
four different random generators provided by rjags.
}
\details{
It is optimal to use four chains. This also allows the sampling to
use the four different samplers implemented in rjags (e.g. Wichmann-Hill,
Marsaglia-Multicarry, Super-Duper, Mersenne-Twister)
}
