% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence_diagnostics.R
\name{get_inits_overdispersed_two_chains}
\alias{get_inits_overdispersed_two_chains}
\title{Helper function to define overdispersed initial values}
\usage{
get_inits_overdispersed_two_chains(
  merging_sections = FALSE,
  ordinal_scale = FALSE,
  point_scale = NULL,
  n_applications,
  n_voters,
  n_sections = NULL,
  grades = NULL,
  seed
)
}
\arguments{
\item{merging_sections}{should the sections be merged?, default is FALSE,
e.g. a ranking is done for each section/panel separately.}

\item{ordinal_scale}{are the evaluation scores on an ordinal scale?,
default is FALSE.}

\item{point_scale}{if on an ordinal scale, what is the number of points on
the scale?, default is NULL}

\item{n_applications}{number of applications / proposals considered.}

\item{n_voters}{number of voters / panel members.}

\item{n_sections}{number of sections / panels, if merged, default is NULL.}

\item{grades}{the grades given to the proposals (the whole
n_applications*n_voters long vector)}

\item{seed}{seed used for the sampling}
}
\value{
a list of overdispersed initial values for the two chains
}
\description{
(note that this function can only be applied when two chains are used,
otherwise the user has to provide the initial values)
}
