% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence_diagnostics.R
\name{get_max_rhat}
\alias{get_max_rhat}
\title{Function to get the maximal values of the Rhat statistic}
\usage{
get_max_rhat(
  data,
  id_application,
  id_voter,
  grade_variable,
  n_burnin = 1000,
  n_iter = 5000,
  n_adapt = 1000,
  n_chains = 2,
  inits_type = "random",
  id_section = NULL,
  tau_section_name = NULL,
  other_variables = c("nu", "voter_intercept"),
  ordinal_scale = FALSE,
  point_scale = NULL,
  path_to_jags_model = NULL,
  initial_values = NULL,
  seed = 1991
)
}
\arguments{
\item{data}{long data frame with all information as in the jags model
defined below.}

\item{id_application}{the name of the application variable in the data}

\item{id_voter}{the name of the voter variable in the data (default = NULL)}

\item{grade_variable}{the name of the outcome variable}

\item{n_burnin}{number of burnin iterations discarded. (default = 1000)}

\item{n_iter}{how many iterations used in the JAGS sampler?
(default = 5000)}

\item{n_adapt}{number of iterations discarded for the adaptation phase.
(default = 1000)}

\item{n_chains}{number of chains for the JAGS sampler. (default = 2)}

\item{inits_type}{type of the initial values, default is "random", but if two
chains are used, the initial values can also be  "overdispersed"}

\item{id_section}{name of the section/panel variable}

\item{tau_section_name}{name of the standard error of the section effect, if
needed}

\item{other_variables}{are there other variables we would like to extract?
(NULL by default)}

\item{ordinal_scale}{dummy variable informing us on whether or not the
outcome is on an ordinal scale (default = FALSE)}

\item{point_scale}{integer informing us on the number of points of the
ordinal scale; not needed for continuous scale (default = NULL)}

\item{path_to_jags_model}{the path to the jags txt file, if null, the
default model is used. (default = NULL)}

\item{initial_values}{The list of initial values can be provided directly}

\item{seed}{set a seed for the JAGS model (default = 1991)}
}
\value{
a number: the maximal values of the computed Rhats
}
\description{
(defaults implemented for two chains)
}
