% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_mock_data}
\alias{get_mock_data}
\title{Load mock data}
\usage{
get_mock_data(panels = "all")
}
\arguments{
\item{panels}{default to 'all', for three panels. other possibilities:
'one', 'two', and 'three'.}
}
\description{
This function loads and returns mock data
}
