% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_num_grade_snsf}
\alias{get_num_grade_snsf}
\title{SNSF numeric grade transformation}
\usage{
get_num_grade_snsf(grade)
}
\arguments{
\item{grade}{the grade ("character")}
}
\description{
This function transforms the ordinal grade to a numeric grade.
}
\details{
A = 6, AB = 5, B = 4, BC = 3, C = 2, D = 1. All other grades (COI,
etc) are transformed to NA.
}
