% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{get_sankey_plot_br}
\alias{get_sankey_plot_br}
\title{Sankey plot for funding groups}
\usage{
get_sankey_plot_br(
  get_mcmc_samples_result,
  n_proposals,
  raw_data,
  num_grade_variable = "num_grade",
  id_proposal,
  name_er_or_theta = "rank_theta",
  er = TRUE,
  title = NULL,
  number_fundable = NULL,
  colors = NULL
)
}
\arguments{
\item{get_mcmc_samples_result}{output from `get_mcmc_sample()`}

\item{n_proposals}{number of proposals in the panel / call}

\item{raw_data}{raw long data with individual votes and numeric grade}

\item{num_grade_variable}{name of numeric grade in raw_data}

\item{id_proposal}{name of proposal / applicant identifier}

\item{name_er_or_theta}{the name of the parameter estimating the ranks (or
the proposal effect) in the Bayesian Hierarchical model
(default = "rank_theta").}

\item{er}{if TRUE then, the ER are plotted, otherwise the thetas (the
proposal effects) are plotted. (default = TRUE)}

\item{title}{title of the plot (default = NULL, no title).}

\item{number_fundable}{number of proposals that can be funded.}

\item{colors}{vector with colors, default = color blind.}
}
\value{
A plot (similar to Sankey plot)
}
\description{
Sankey plot for funding groups
}
