% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/er_and_sucra.R
\name{get_sucra}
\alias{get_sucra}
\title{The Surface Under the Cumulative RAnking (SUCRA)}
\usage{
get_sucra(
  data,
  id_application,
  id_voter,
  grade_variable,
  path_to_jags_model = NULL,
  n_chains = 3,
  n_iter = 5000,
  n_burnin = 1000,
  id_section = NULL,
  theta_name = "application_intercept",
  tau_name = "tau_application",
  tau_voter_name = "tau_voter",
  tau_section_name = NULL,
  sigma_name = "sigma",
  other_variables = NULL,
  rank_theta_name = "rank_theta",
  mcmc_samples = NULL,
  seed = 1991
)
}
\arguments{
\item{data}{long data frame with all information as in the jags model
defined below.}

\item{id_application}{the name of the application variable in the data}

\item{id_voter}{the name of the voter variable in the data (default = NULL)}

\item{grade_variable}{the name of the outcome variable}

\item{path_to_jags_model}{the path to the jags txt file, if null, the
default model is used. (default = NULL)}

\item{n_chains}{number of chains for the JAGS sampler. (default = 4)}

\item{n_iter}{how many iterations used in the JAGS sampler? (default = 5000)}

\item{n_burnin}{number of burnin iterations}

\item{id_section}{name of the section}

\item{theta_name}{the name of the application identifier. (default =
application_intercept")}

\item{tau_name}{name of the the sqrt of tau, being the precision of the
random effects, in the jags model. (default = sd_application)}

\item{tau_voter_name}{name of the (default = tau_voter)}

\item{tau_section_name}{name of the standard error of the section effect, if
needed (default = NULL).}

\item{sigma_name}{name of the standard deviation of the full model.
(default = sigma)}

\item{other_variables}{are there other variables we would like to extract?
(NULL by default)}

\item{rank_theta_name}{the name of the rank of theta in the JAGS model
(default = rank_theta)}

\item{mcmc_samples}{if the mcmc sample has already been run (default = NULL).}

\item{seed}{set a seed for the JAGS model (default = 1991)}
}
\value{
the result is a names vector with the SUCRA of all applications
}
\description{
This function computes the SUCRA
}
\examples{
data_panel1 <- get_mock_data() \%>\%
         filter(panel == "p1")
\dontrun{
SUCRA_results <- get_sucra(data = data_panel1,
                           id_application = "application",
                           id_voter = "voter", grade_variable = "num_grade")
                           }
}
