% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_er_distributions}
\alias{plot_er_distributions}
\title{Expected Ranks with Credible intervals}
\usage{
plot_er_distributions(
  get_mcmc_samples_result,
  n_proposals,
  name_er_or_theta = "rank_theta",
  er = TRUE,
  title = NULL,
  names_proposals = NULL,
  inner_credible = 0.5,
  outer_credible = 0.9,
  ylab = "Expected Ranks",
  number_fundable = NULL,
  outer_show = TRUE,
  size_pt = 1.5,
  size_outer = 0.5,
  size_inner = 2,
  alpha_inner = 0.5,
  proposal = "Proposal "
)
}
\arguments{
\item{get_mcmc_samples_result}{the mcmc samples}

\item{n_proposals}{number of proposals in the panel / call}

\item{name_er_or_theta}{the name of the parameter estimating the ranks (or
the proposal effect) in the Bayesian Hierarchical model
(default = "rank_thetta")}

\item{er}{if TRUE then, the ER are plotted, otherwise the thetas (the
proposal effects) are plotted. (default = TRUE)}

\item{title}{title of the plot (default = NULL, no title).}

\item{names_proposals}{names of the proposals to be written on the x-axis
ticks. Note that the ordering has to be the same as the one they are given to
the `get_mcmc_samples()` function (default = NULL, they are simply numerated).}

\item{inner_credible}{percentage of the inner credible interval (default = .5)}

\item{outer_credible}{percentage of the outer credible interval (default = .9)}

\item{ylab}{The y axis label (default = "Expected Ranks")
(default = NULL, no FL drawn).}

\item{number_fundable}{number of proposals that can be funded}

\item{outer_show}{should the outer credible interval be shown (default = TRUE)}

\item{size_pt}{the size of the point estimate (mean, ER) (default = 1.5)}

\item{size_outer}{the size of the line representing the outer CrI
(default = .5)}

\item{size_inner}{the size of the line representing the inner CrI
(default = 2)}

\item{alpha_inner}{alpha for the color of the inner CrI}

\item{proposal}{Word written infront of number (default = "Proposal")}
}
\value{
the result is a plot of the expected rank together with their
credible intervals
}
\description{
This function plots the expected ranks of all proposals together with
their xy% credible intervals.
}
