% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_rankogram}
\alias{plot_rankogram}
\title{Rankogram}
\usage{
plot_rankogram(
  data,
  id_application,
  id_voter,
  grade_variable = "num_grade",
  path_to_jags_model = NULL,
  n_chains = 3,
  n_iter = 5000,
  n_burnin = 1000,
  id_section = NULL,
  rank_theta_name = "rank_theta",
  theta_name = "application_intercept",
  tau_name = "tau_application",
  tau_voter_name = "tau_voter",
  sigma_name = "sigma",
  cumulative_rank_prob = FALSE,
  mcmc_samples = NULL
)
}
\arguments{
\item{data}{long data frame with all information as in the jags model
defined below.}

\item{id_application}{the name of the application variable in the data}

\item{id_voter}{the name of the voter variable in the data (default = NULL)}

\item{grade_variable}{the name of the outcome variable
(default = "num_grade")}

\item{path_to_jags_model}{the path to the jags txt file, if null, the
default model is used. (default = NULL)}

\item{n_chains}{number of chains for the JAGS sampler. (default = 4)}

\item{n_iter}{how many iterations used in the JAGS sampler? (default = 5000)}

\item{n_burnin}{number of burnin iterations}

\item{id_section}{name of the section}

\item{rank_theta_name}{the name of the rank of theta in the Bayesian model.
(default = rank_theta")}

\item{theta_name}{the name of the application identifier. (default =
"application_intercept")}

\item{tau_name}{name of the the squareroot of tau, being the precision of the
random effects, in the jags model. (default = sd_application)}

\item{tau_voter_name}{name of the (default = tau_voter)}

\item{sigma_name}{name of the standard error of th term (default = sigma)}

\item{cumulative_rank_prob}{should the cumulative ranking probabilities be
represented instead}

\item{mcmc_samples}{if the mcmc sample has already been run (default = NULL).}
}
\value{
the result is a plot of the rankogram (or cumulative ranking
probabilities)
}
\description{
This function produces a rankogram
}
