% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plotting_er_results}
\alias{plotting_er_results}
\title{The Plot of the Expected Rank from a JAGS model}
\usage{
plotting_er_results(
  er_results,
  rankability = TRUE,
  id_application = "id_application",
  how_many_fundable = NULL,
  title = "",
  ordering_increasing = TRUE,
  draw_funding_line = TRUE,
  result_show = TRUE,
  easy_numbering = TRUE,
  colors = NULL,
  pt_size = 1,
  line_size = 0.5,
  alpha_line = 1,
  alpha_pt = 1,
  line_type_fl = "longdash",
  x_expand = 0.6,
  alpha_fl = 0.5,
  color_fl = "darkgray",
  size_fl = 2,
  padding_right = 0.3,
  padding_left = 0.1,
  grep_size = 2,
  nudge_x = 0.5,
  min_segment_length = 0,
  segment_curvature = -0.1,
  segment_angle = 20,
  segment_ncp = 3,
  segment_size = 0.15,
  hjust_y_axis = 1.5
)
}
\arguments{
\item{er_results}{the resulting object from the `get_er_from_jags()` function}

\item{rankability}{Do we want the rankability written in the plot title
(default = TRUE)}

\item{id_application}{the name of the variable giving the distinct
application identifiers. (default = id_application)}

\item{how_many_fundable}{number of proposals that can be funded?
(default = NULL) If Null, the graph will be black and white only}

\item{title}{beginning of the plot title (default = " ").}

\item{ordering_increasing}{should the order of the rank be increasing, from
up to dowm: from rank 1 on top to rank n on the bottom (default = TRUE)}

\item{draw_funding_line}{should the funding line be drawn? (default = TRUE)}

\item{result_show}{boolean telling us whether the application number/id
should be added next to the ER points (default = FALSE).}

\item{easy_numbering}{should the ER results just be numbered?
(default = TRUE)}

\item{colors}{a color pallette that can be changes.}

\item{pt_size}{size of points (default = 1)}

\item{line_size}{size of the lines (default = 0.5)}

\item{alpha_line}{...}

\item{alpha_pt}{...}

\item{line_type_fl}{... fl = funding line}

\item{x_expand}{make funding line longer on both sides (default = .3)}

\item{alpha_fl}{...}

\item{color_fl}{...}

\item{size_fl}{...}

\item{padding_right}{for ggrepel}

\item{padding_left}{for ggrepel}

\item{grep_size}{the size of the labels on the points (default = .25).}

\item{nudge_x}{element of ggrepel.}

\item{min_segment_length}{...}

\item{segment_curvature}{...}

\item{segment_angle}{...}

\item{segment_ncp}{...}

\item{segment_size}{...}

\item{hjust_y_axis}{...}
}
\value{
the result is a plot
}
\description{
This function plot the results form the JAGS-ER
}
\examples{
dat <- get_mock_data() \%>\%
filter(panel == "p1")
ER_results <- get_er_from_jags(data = dat,
                  id_application = "application",
                  id_voter = "voter",
                  grade_variable = "num_grade",
                  n_chains = 2, n_iter = 1000,
                  n_burnin = 1000)
plotting_er_results(ER_results, title = "Panel 1", how_many_fundable = 5)

}
