% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plotting_er_with_var}
\alias{plotting_er_with_var}
\title{The Plot of the Expected Rank with their confidence intervals}
\usage{
plotting_er_with_var(
  er_results,
  id_application = "id_application",
  alpha = 0.5,
  color = 1
)
}
\arguments{
\item{er_results}{the resulting object from the `get_er_from_jags()` function}

\item{id_application}{name of the identifier of the application}

\item{alpha}{alpha level for the CI-bands (default = 0.5)}

\item{color}{color of point and bands (default = "black")}
}
\description{
The ER of each application are presented together with their 95% Wald CI
}
