% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jags_functions.R
\name{get_default_jags_model}
\alias{get_default_jags_model}
\title{Write a default jags model}
\usage{
get_default_jags_model(path = "default_jags_model.txt")
}
\arguments{
\item{path}{where should it be stored, from here()? (default = jags_model.txt)}
}
\description{
This function simply stores a txt file with a default jags model, in a
certain path.
}
\details{
The model defined here has a random component for the application /
the proposal and the voter / the evaluator. There is no other grouping
variable defined, as for example a section or panel. The user is invited to
write their own model definition if more flexibility is needed.
}
\examples{
# The model definition .txt is stored in the file "default_jags_model.txt"
\dontrun{
get_default_jags_model()
}
}
