% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jags_functions.R
\name{get_default_jags_model}
\alias{get_default_jags_model}
\title{Write a default jags model}
\usage{
get_default_jags_model(
  outcome_variable = "continuous",
  residuals = "homogeneous",
  path = "default_jags_model.txt",
  quiet = FALSE
)
}
\arguments{
\item{outcome_variable}{the nature of the outcome variable: continuous or
ordinal. (default = "continuous")}

\item{residuals}{the nature of the residuals in the model: homogeneous or
heterogeneous. (default = "homogeneous")}

\item{path}{where should it be stored, from here()? (default = jags_model.txt)}

\item{quiet}{if TRUE, do not show messages.}
}
\description{
This function simply stores a txt file with a default jags model, in a
certain path.
}
\details{
The model defined here has a random component for the application /
the proposal and the voter / the evaluator. There is no other grouping
variable defined, as for example a section or panel. The user is invited to
write their own model definition if more flexibility is needed. However, the
user can decide between a continuous or ordinal outcome variable, and
homogeneous or heterogenous residuals.
}
\examples{
# The model definition .txt is stored in the file "default_jags_model.txt"
\dontrun{
get_default_jags_model()
}
}
