% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_right_data_format}
\alias{get_right_data_format}
\title{Individual votes to correct format}
\usage{
get_right_data_format(
  individual_votes,
  prefix_voter = "voter",
  fun_grade_to_num = get_num_grade_snsf,
  delete_NA = TRUE
)
}
\arguments{
\item{individual_votes}{The matrix containing the individual votes.}

\item{prefix_voter}{The column name prefix used to indicate the voters.}

\item{fun_grade_to_num}{function that translates grades to numeric values}

\item{delete_NA}{should the missing (numeric) grades be deleted?
(default = TRUE)}
}
\description{
This function transforms the default csv file with the individual votes into
the correct long format
}
